fix(client): remove react native tabs
#24
