+++ packages/client/package.json
@@ -28,7 +28,6 @@
     "react-native-reanimated": "~1.9.0",
     "react-native-safe-area-context": "~3.0.7",
     "react-native-screens": "~2.9.0",
-    "react-native-tab-view": "^2.14.4",
     "react-native-web": "~0.11.7",
     "recoil": "^0.0.10",
     "socket.io-client": "^2.3.0"
