+++ packages/client/src/screens/Game.tsx
@@ -4,25 +4,34 @@ import { useTheme } from 'react-native-paper';
 
 import GameAnswer from '../components/GameAnswer';
 import GameInput from '../components/GameInput';
-import GameQuestion from '../components/GameQuestion';
-import { LargeScoreBoard } from '../components/ScoreBoard';
-import { useSocketListener } from '../utils/hooks/socketListener';
+import { ScoreBoard } from '../components/ScoreBoard';
+import { useScreenWidth } from '../utils/hooks/screenWidth';
 
 export default function Game() {
   const { colors } = useTheme();
-  const question = useSocketListener('question', null);
+  const isLargeScreen = useScreenWidth();
+
   return (
-    <View style={[styles.container, { backgroundColor: colors.primary }]}>
-      <LargeScoreBoard />
-      <View style={styles.gameContainer}>
-        <View style={styles.info}>
-          <GameAnswer />
-        </View>
-        <View style={styles.info}>
-          <GameQuestion question={question} />
-        </View>
-        <GameInput question={question} />
-      </View>
+    <View
+      style={[
+        styles.container,
+        { backgroundColor: colors.primary },
+        isLargeScreen && { flexDirection: 'row' },
+      ]}>
+      {!isLargeScreen && <GameAnswer />}
+      <ScoreBoard />
+      {isLargeScreen ? (
+        <>
+          <View style={styles.gameContainer}>
+            <View style={styles.info}>
+              <GameAnswer />
+            </View>
+            <GameInput />
+          </View>
+        </>
+      ) : (
+        <GameInput />
+      )}
     </View>
   );
 }
@@ -30,7 +39,6 @@ export default function Game() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    flexDirection: 'row',
   },
   gameContainer: {
     flex: 1,
